Add script to download original images and video

+++ parser.py
@@ -64,7 +64,8 @@ def tweet_json_to_markdown(tweet, username, archive_media_folder, output_media_f
                 new_url = output_media_folder_name + tweet_id_str + '-' + original_filename
                 if os.path.isfile(local_filename):
                     # Found a matching image, use this one
-                    shutil.copy(local_filename, new_url)
+                    if not os.path.isfile(new_url):
+                        shutil.copy(local_filename, new_url)
                     markdown = f'![]({new_url})'
                 else:
                     # Is there any other file that includes the tweet_id in its filename?
@@ -73,7 +74,8 @@ def tweet_json_to_markdown(tweet, username, archive_media_folder, output_media_f
                         markdown = ''
                         for media_filename in media_filenames:
                             media_url = f'{output_media_folder_name}{os.path.split(media_filename)[-1]}'
-                            shutil.copy(media_filename, media_url)
+                            if not os.path.isfile(media_url):
+                                shutil.copy(media_filename, media_url)
                             markdown += f'\n\n<video controls><source src="{media_url}">Your browser does not support the video tag.</video>\n{media_url}'
                     else:
                         print(f'Warning: missing local file: {local_filename}. Using original link instead: {original_url} (expands to {original_expanded_url})')
@@ -138,6 +140,10 @@ def main():
         with open(filename, 'w', encoding='utf-8') as f:
             f.write(md)
     print(f'Wrote to tweets_YYYY-MM.md, with images and video embedded from {output_media_folder_name}')
+
+    # Tell the user that it is possible to download better-quality media
+    print("\nThe archive doesn't contain the original-size images. If you are interested in retrieving the original images")
+    print("from Twitter then please run the script download_better_images.py")
 
 
 if __name__ == "__main__":
